graphyne-ai/requirements.txt
agno
pydantic>=2.0
openai
requests
python-dotenv